pyproject.toml
[project]
name = "semester-project"
version = "0.1.0"
description = ""
authors = [
    {name = "thediogen",email = "evgendvodnenko@gmail.com"}
]
readme = "README.md"
requires-python = "3.11.11"
dependencies = [
    "fastapi (>=0.115.6,<0.116.0)",
    "pydantic (>=2.10.5,<3.0.0)",
    "aioredis (>=2.0.1,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "greenlet (>=3.1.1,<4.0.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "redis (>=5.2.1,<6.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (>=4.2.1,<5.0.0)",
    "itsdangerous (>=2.2.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
